move sensorStorage stuff out of sigksens.h
