+++ SigkSens/SigkSens.ino
@@ -2,6 +2,7 @@
 
 
 #include "config.h"
+
 #ifdef ENABLE_I2C
 #include "src/services/i2c.h"
 #endif
@@ -45,6 +46,7 @@
 
 
 #include "sigksens.h"
+#include "src/sensors/sensorStorage.h"
 
 /*---------------------------------------------------------------------------------------------------
 -----------------------------------------------------------------------------------------------------
