+++ SigkSens/src/sensors/ads1115/ads1115.h
@@ -2,6 +2,7 @@
 #define _ads1115_H_
 
 #include "../../../sigksens.h"
+#include "../sensorStorage.h"
