+++ SigkSens/src/sensors/sensofInfo.cpp
@@ -1,0 +1,8 @@
+
+#include <ArduinoJson.h>
+
+
+#include "sensorInfo.h"
+
+
+fromJsonFunc fromJson[(int)SensorType::SensorType_MAX+1];
